Release/0.1.0 (#8)
* Set pump initial settings and pump popover

* Check reservoir

* fix initial setup

* open CGM source

* fix initial pump setup

* Hide debug options

* fix initial preferences

* use xcconfig

* Autoupdate build number

* Create README_RU.md

* Display version and build number

* English README

* Disclaimer

* disallow bolus 0

* possible fix wrong layout after start

* fix glucose Y range

* Optional api secret

* fix suggestion parsing

* Face id access

* fix clock

* build autoupdate

* Fix temp targets layout and pump view

* Fix pump updating logic

* fix omnipod reservooir

* Background timers

* legend

* Remove unused dependecies

* show temp target range

* FAQ_RU

* FAQ eng

* bump buils number

* clear signing

+++ FreeAPS/Sources/Models/Suggestion.swift
@@ -14,7 +14,7 @@ struct Suggestion: JSON, Equatable {
     let deliverAt: Date?
     let carbsReq: Decimal?
     let temp: TempType?
-    let bg: Int?
+    let bg: Decimal?
     let reservoir: Decimal?
     var timestamp: Date?
     var recieved: Bool?
